backend/requirements.txt
# Backend Service Requirements(production)

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# CORS
starlette==0.35.1

# Environment
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
orjson==3.9.12
